+++ src/react-media-player.js
@@ -1,6 +1,5 @@
 export Media from './Media'
 export Player from './Player'
 export withMediaProps from './decorators/with-media-props'
-export withKeyboardControls from './decorators/with-keyboard-controls'
 export * as controls from './controls/exports'
 export * as utils from './utils/exports'