+++ src/utils/exports.js
@@ -1,1 +1,2 @@
+export keyboardControls from './keyboard-controls'
 export formatTime from './format-time'